+++ NEWS.md
@@ -1,7 +1,9 @@
 # moodef (development version)
 
-# moodef 1.0.0
+* Generate questions from an Excel file. 
+* Create an Excel file.
+* Read an Excel file.
 
-# moodef 0.1.0
+# moodef 1.0.0
 
 * Initial CRAN submission.